+++ compose/foundation/foundation/src/uikitMain/kotlin/androidx/compose/foundation/gestures/FlingBehavior.uikit.kt
@@ -18,6 +18,7 @@ package androidx.compose.foundation.gestures
 
 import androidx.compose.animation.core.cupertino.CupertinoScrollDecayAnimationSpec
 import androidx.compose.animation.core.generateDecayAnimationSpec
+import androidx.compose.foundation.gestures.cupertino.CupertinoFlingBehavior
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.platform.LocalDensity
